// src/web/types.ts
export enum PlayerEvent {
  AdBreakFinished = "adbreakfinished",
  AdBreakStarted = "adbreakstarted",
  AdClicked = "adclicked",
  AdError = "aderror",
  AdFinished = "adfinished",
  AdInteraction = "adinteraction",
  AdLinearityChanged = "adlinearitychanged",
  AdManifestLoaded = "admanifestloaded",
  AdQuartile = "adquartile",
  AdSkipped = "adskipped",
  AdStarted = "adstarted",
  AirplayAvailable = "airplayavailable",
  AirplayChanged = "airplaychanged",
  AspectRatioChanged = "aspectratiochanged",
  AudioAdaptation = "audioadaptation",
  AudioAdded = "audioadded",
  AudioChanged = "audiochanged",
  AudioDownloadQualityChange = "audiodownloadqualitychange",
  AudioDownloadQualityChanged = "audiodownloadqualitychanged",
  AudioPlaybackQualityChanged = "audioplaybackqualitychanged",
  AudioQualityAdded = "audioqualityadded",
  AudioQualityChanged = "audioqualitychanged",
  AudioQualityRemoved = "audioqualityremoved",
  AudioRemoved = "audioremoved",
  CastAvailable = "castavailable",
  CastStart = "caststart",
  CastStarted = "caststarted",
  CastStopped = "caststopped",
  CastWaitingForDevice = "castwaitingfordevice",
  ContentLocationChanged = "contentlocationchanged",
  CueEnter = "cueenter",
  CueExit = "cueexit",
  CueParsed = "cueparsed",
  CueUpdate = "cueupdate",
  DVRWindowExceeded = "dvrwindowexceeded",
  Destroy = "destroy",
  DownloadFinished = "downloadfinished",
  DrmLicenseAdded = "drmlicenseadded",
  DurationChanged = "durationchanged",
  Error = "error",
  LatencyModeChanged = "latencymodechanged",
  LicenseValidated = "licensevalidated",
  Metadata = "metadata",
  MetadataChanged = "metadatachanged",
  MetadataParsed = "metadataparsed",
  ModuleReady = "moduleready",
  Muted = "muted",
  OverlayAdStarted = "overlayadstarted",
  Paused = "paused",
  PeriodSwitch = "periodswitch",
  PeriodSwitched = "periodswitched",
  Play = "play",
  PlaybackFinished = "playbackfinished",
  PlaybackSpeedChanged = "playbackspeedchanged",
  PlayerResized = "playerresized",
  Playing = "playing",
  Ready = "ready",
  RestoringContent = "restoringcontent",
  Seek = "seek",
  Seeked = "seeked",
  SegmentPlayback = "segmentplayback",
  SegmentRequestFinished = "segmentrequestfinished",
  ShowAirplayTargetPicker = "showairplaytargetpicker",
  SourceLoaded = "sourceloaded",
  SourceUnloaded = "sourceunloaded",
  StallEnded = "stallended",
  StallStarted = "stallstarted",
  SubtitleAdded = "subtitleadded",
  SubtitleDisable = "subtitledisable",
  SubtitleDisabled = "subtitledisabled",
  SubtitleEnable = "subtitleenable",
  SubtitleEnabled = "subtitleenabled",
  SubtitleRemoved = "subtitleremoved",
  TargetLatencyChanged = "targetlatencychanged",
  TimeChanged = "timechanged",
  TimeShift = "timeshift",
  TimeShifted = "timeshifted",
  Unmuted = "unmuted",
  VRStereoChanged = "vrstereochanged",
  VRViewingDirectionChange = "vrviewingdirectionchange",
  VRViewingDirectionChanged = "vrviewingdirectionchanged",
  VideoAdaptation = "videoadaptation",
  VideoDownloadQualityChange = "videodownloadqualitychange",
  VideoDownloadQualityChanged = "videodownloadqualitychanged",
  VideoPlaybackQualityChanged = "videoplaybackqualitychanged",
  VideoQualityAdded = "videoqualityadded",
  VideoQualityChanged = "videoqualitychanged",
  VideoQualityRemoved = "videoqualityremoved",
  ViewModeChanged = "viewmodechanged",
  VolumeChanged = "volumechanged",
  Warning = "warning",
}